table-border-spacing.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS test: calc() and border-spacing</title>
<meta name="assert" content="calc() can be used in the border-spacing property">
<meta name="flags" content="">
<link rel="help" href="https://drafts.csswg.org/css-values-3/#calc-notation">
<link rel="help" href="https://drafts.csswg.org/css2/tables.html#propdef-border-spacing">
<script src="/resources/testharness.js" ></script>
<script src="/resources/testharnessreport.js"></script>
<style>
    td { 
        padding: 0; 
        width: 50px; 
        height: 60px; 
        background-color: blue; 
        border: 1px solid black 
    }
    table { 
        border-spacing: calc(10px * 2 + 5px); 
        border: 3px solid black; 
    }
</style>
<table id="stimulus">
    <tr>
        <td></td>
        <td></td>
        <td></td>
        <td></td>
        <td></td>
    </tr>
    <tr>
        <td></td>
        <td></td>
        <td></td>
        <td></td>
        <td></td>
    </tr>
</table>
<script>
test(function() {
    var innerBorder = 1;
    var outerBorder = 3;
    var innerWidth = 50 + 2 * innerBorder;
    var innerHeight = 60 + 2 * innerBorder;
    var borderSpacing = 25;
    var rows = 2;
    var columns = 5;

    var expectedWidth = innerWidth * columns + borderSpacing * (columns + 1) + 2 * outerBorder;
    var expectedHeight = innerHeight * rows + borderSpacing * (rows + 1) + 2 * outerBorder;

    assert_equals(document.getElementById("stimulus").offsetWidth, expectedWidth, "the width is as expected");
    assert_equals(document.getElementById("stimulus").offsetHeight, expectedHeight, "the height is as expected");
}, "calc() in border-spacing");
</script>    
